+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Next version
+
+### 🚀 New
+
+* Add flag `--offset-min-skybrightness` to `jaeger configuration load/preload` to set the `coordio.utils.offset_definition` `offset_min_skybrightness` parameters. Defaults to 0.5.
+
+
 ## 1.3.3 - January 15, 2023
 
 ### ✨ Improved